User management theming and consistency
Added tables and theming to each of the user management forms.
Changed references to email addresses for users in user management to
username to fit what is being queried and used.
Deleted password.php and registerOld.php files as they are no longer
needed in the project.
Added a gradient background to entire web site to give a better feeling
user experience.